requirements.txt
discord.py>=2.3.2
gpxpy>=1.6.2
aiosqlite>=0.20.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
